+++ Freshmart.Models/Queries/Product/GetAllProductQuery.cs
@@ -11,7 +11,7 @@ namespace FreshMart.Models.Queries
     {
         public string Text { get; set; }
 
-        public int? DistrictId { get; set; }
+        public long? DistrictId { get; set; }
 
         public List<int> PriceRanges { get; set; }
